# openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Jtl Reporter Api",
    "description" : "Jtl Reporter Api",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
    "/projects" : {
      "get" : {
        "tags" : [ "projects" ],
        "summary" : "returns list of projects",
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectItem"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "projects" ],
        "summary" : "create new project",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewProject"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/projects/latest-items" : {
      "get" : {
        "tags" : [ "projects" ],
        "summary" : "get latest project items",
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectLatestItems"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          }
        }
      }
    },
...
    "/projects/{projectName}" : {
      "put" : {
        "tags" : [ "projects" ],
        "summary" : "updates project name",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewProject"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "No content"
          }
        }
      },
      "delete" : {
        "tags" : [ "projects" ],
        "summary" : "deletes project",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No content"
          }
        }
      }
    },
...
    "/projects/{projectName}/scenarios/{scenarioName}/items" : {
      "get" : {
        "tags" : [ "items" ],
        "summary" : "returns list of items",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scenarioName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "The number of items to skip before starting to collect the result set",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The numbers of items to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Items"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found"
          }
        }
      },
      "post" : {
        "tags" : [ "items" ],
        "summary" : "create new item",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scenarioName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewItem"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ItemSave"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          }
        }
      }
    },
    "/projects/{projectName}/scenarios/{scenarioName}/items/{itemId}" : {
      "get" : {
        "tags" : [ "items" ],
        "summary" : "returns item detail",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scenarioName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Item"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "items" ],
        "summary" : "updates item",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scenarioName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "No content"
          },
          "400" : {
            "description" : "Bad request"
          }
        }
      },
      "delete" : {
        "tags" : [ "items" ],
        "summary" : "deletes item",
        "parameters" : [ {
          "name" : "projectName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scenarioName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No content"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ProjectItem" : {
        "required" : [ "id", "itemCount", "projectName", "scenarioCount" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "projectName" : {
            "type" : "string",
            "example" : "My Project"
          },
          "itemCount" : {
            "type" : "number",
            "example" : 10
          },
          "scenariocount" : {
            "type" : "number",
            "example" : 3
          },
          "latestRun" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "NewProject" : {
        "required" : [ "projectName" ],
        "type" : "object",
        "properties" : {
          "projectName" : {
            "maxLength" : 50,
            "minLength" : 3,
            "type" : "string",
            "example" : "My Project"
          }
        }
      },
      "ProjectLatestItems" : {
        "required" : [ "id", "projectName" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "name" : {
            "type" : "string"
          },
          "environment" : {
            "type" : "string"
          },
          "projectName" : {
            "type" : "string"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string",
            "example" : "passed"
          }
        }
      },
...
      "ScenarioData" : {
        "required" : [ "avgLatency", "avgResponseTime", "duration", "endDate", "errorRate", "maxVu", "percentil", "startDate", "throughput" ],
        "type" : "object",
        "properties" : {
          "avgLatency" : {
            "type" : "number",
            "example" : 32
          },
          "avgResponseTime" : {
            "type" : "number",
            "example" : 86
          },
          "duration" : {
            "type" : "number",
            "format" : "float",
            "example" : 1.4
          },
          "errorRate" : {
            "type" : "number",
            "example" : 1
          },
          "maxVu" : {
            "type" : "number",
            "example" : 50
          },
          "percentil" : {
            "type" : "number",
            "example" : 39
          },
          "throughput" : {
            "type" : "number",
            "format" : "float",
            "example" : 1.6
          },
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "Items" : {
        "required" : [ "data", "name", "total" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "scenario name"
          },
          "total" : {
            "type" : "number",
            "example" : 20
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ItemsData"
            }
          }
        }
      },
      "ItemsData" : {
        "required" : [ "base", "duration", "id", "maxVu", "startTime", "status", "uploadTime" ],
        "type" : "object",
        "properties" : {
          "base" : {
            "type" : "boolean"
          },
          "duration" : {
            "type" : "number",
            "format" : "float",
            "example" : 10.1
          },
          "environment" : {
            "type" : "string",
            "example" : "prod"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "maxVu" : {
            "type" : "number",
            "example" : 30
          },
          "note" : {
            "type" : "string",
            "example" : "test note"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "example" : "1"
          },
          "uploadTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "NewItem" : {
        "type" : "object",
        "properties" : {
          "environment" : {
            "type" : "string",
            "example" : "test"
          },
          "note" : {
            "type" : "string",
            "example" : "test note"
          },
          "status" : {
            "type" : "string",
            "example" : "1",
            "enum" : [ "10", "0", "1", "2", "3" ]
          },
          "kpi" : {
            "format" : "binary"
          },
          "errors" : {
            "format" : "binary"
          }
        }
      },
      "ItemSave" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "overview" : {
            "$ref" : "#/components/schemas/ScenarioData"
          },
          "status" : {
            "type" : "string",
            "example" : "1",
            "enum" : [ "10", "0", "1", "2", "3" ]
          }
        }
      },
      "Item" : {
        "type" : "object",
        "properties" : {
          "isBase" : {
            "type" : "boolean"
          },
          "baseId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "environment" : {
            "type" : "string",
            "example" : "test"
          },
          "note" : {
            "type" : "string",
            "example" : "test note"
          },
          "overview" : {
            "$ref" : "#/components/schemas/ScenarioData"
          },
          "status" : {
            "type" : "string",
            "example" : "1",
            "enum" : [ "10", "0", "1", "2", "3" ]
          },
          "statistics" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Statistics"
            }
          },
          "plot" : {
            "$ref" : "#/components/schemas/Item_plot"
          }
        }
      },
      "Statistics" : {
        "type" : "object",
        "properties" : {
          "avgResponseTime" : {
            "type" : "number",
            "example" : 30
          },
          "errorRate" : {
            "type" : "number",
            "format" : "float",
            "example" : 4.3
          },
          "label" : {
            "type" : "string",
            "example" : "label name"
          },
          "maxResponseTime" : {
            "type" : "number",
            "example" : 432
          },
          "minResponseTime" : {
            "type" : "number",
            "example" : 32
          },
          "n0" : {
            "type" : "number",
            "example" : 12
          },
          "n5" : {
            "type" : "number",
            "example" : 21
          },
          "n9" : {
            "type" : "number",
            "example" : 23
          },
          "samples" : {
            "type" : "number",
            "example" : 642
          },
          "throughput" : {
            "type" : "number",
            "format" : "float",
            "example" : 4.2
          }
        }
      },
      "PlotItem" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "name"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number"
              }
            }
          }
        }
      },
      "ItemUpdate" : {
        "required" : [ "base", "environment" ],
        "type" : "object",
        "properties" : {
          "environment" : {
            "maxLength" : 50,
            "minLength" : 1,
            "type" : "string",
            "example" : "test"
          },
          "note" : {
            "maxLength" : 250,
            "type" : "string",
            "example" : "test note"
          },
          "base" : {
            "type" : "boolean"
          }
        }
      },
      "Item_plot" : {
        "type" : "object",
        "properties" : {
          "overallFailRate" : {
            "$ref" : "#/components/schemas/PlotItem"
          },
          "overallThroughput" : {
            "$ref" : "#/components/schemas/PlotItem"
          },
          "overallTimeResponse" : {
            "$ref" : "#/components/schemas/PlotItem"
          },
          "responseTime" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PlotItem"
            }
          },
          "threads" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "number"
              }
            }
          },
          "throughput" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PlotItem"
            }
          }
        }
      }
    }
  }
}